Update API reference for os-floating-ip-pools
The os-floating-ip-pools API is implemented for both
nova-network and neutron. When using neutron, it justs
lists external network names.

Since it's a deprecated proxy to neutron we won't remove
it as part of blueprint remove-nova-network, but we can
update the API reference to explain how to achieve the
equivalent functionality in the networking service API,
so that's done here.

Change-Id: Ieb4c7a0657929f66ed5b642c5941e65aee24372d