[placement] demonstrate part of bug 1778591 with a gabbi test
The test creates a resource provider with a small amount of VCPU
inventory. It then tries to allocate a large number VCPU via a new
consumer (and thus using consumer_generation of null).

This fails with a 409 because capacity is violated.

Then it queries to GET allocations on that consumer, to see if consumer
generation is available. It is not.

The it tries to allocate a smaller amount and though it might be
expected that the generation should still be null, it takes a 0
for the allocation to be accepted.

Once we decide what the right behavior is here, we can fix it.

Change-Id: I196834efc2cb6ece4204abbf8544c8f2621aaa36
Related-Bug: #1778591