Remove unused DB API instance_group_delete method
Since Ifbd53b13fa0fef62e0329283b73d587f367e46c2 we no
longer have compat code in the InstanceGroup object
and everything is handled in the API database, so we can
drop the legacy DB API methods for instance groups.

This is part of a series of cleanup changes.

Change-Id: I569ee5d582ce42d84afe2b4b6e702657abe8733b