Fix unit test modifying global state
The ConductorTaskTestCase.test_build_instances test case uses the
RequestSpec class instead of an instance of the RequestSpec as a mock
return value. If this return value is modified during the test that
modification is done on the class and therefore it is visible to any
subsequent test cases running in the same executor. This could lead to
very strange unit test failures.

This patch adds the missing instantiation to the test case.

Change-Id: Iec47a2c12f35bf050852fe728417c9ddff329a44
(cherry picked from commit 6261325dce258b137ac6013433a50a3e9250a98f)