Wait for vif plugging during live migration job
The nova-live-migration CI job uses OVS which sends
vif plugged events to the source host during live
migration when the destination host plugs vifs
during pre_live_migration. We can set the
live_migration_wait_for_vif_plug option to True
so that the source host waits for the vif to be
plugged and get the event from neutron before
starting the actual guest transfer in the hypervisor.

This is really just about coverage of this code
in a live environment so we don't regress it.

Change-Id: I3d7770589a1526e851d5764214c26ec1a9a8fd43