Use ironic-tempest-dsvm-ipa-wholedisk-bios-agent_ipmitool-tinyipa in tree
ironic-tempest-dsvm-ipa-wholedisk-bios-agent_ipmitool-tinyipa is
defined in the ironic repo but nova uses it because it's defined
in the project-config repo with it's own irrelevant-files list.
Due to a recent change in zuul [1], the definition from project-config
overrides the irrelevant-files from the job definition in ironic and
results in us running this job on things like docs-only changes in
nova, which is not awesome.

This change moves the job usage in-tree from project-config and updates
the irrelevant-files list to match nova-dsvm-base.

Since project-config is branchless, we'll have to backport this in-tree
change to nova's stable branches to continue running this job on nova
stable branches once the override is removed from project-config.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-June/131304.html

Conflicts:
      .zuul.yaml

NOTE(mriedem): The conflict is due to not running the nova-cells-v1
job in stable/pike because of bug 1745838.

Change-Id: I9435fe84c8dc8a6f35ffb195ddbf71ca34edb242
Closes-Bug: #1776981
(cherry picked from commit 17b3bae64ca7aa4c1a6d296860cce6c0bfbb38a1)
(cherry picked from commit 1e16b5f9f341cd0d086065e8b513937f7cd56f5a)