Remove file injection from config drive sample docs
File injection is disabled by default (see
[libvirt]/inject_partition) and deprecated
in the API since microversion 2.57. We
don't really want people relying on file
injection, so this removes it from the
config drive docs. While in here, also
replace the 'nova boot' command mention
with 'openstack server create' since that
is the command used through the examples.

Change-Id: I4521eca6f6cbd4f8cd6dad48ab1d40a1b136bf73