manage: Remove dead code
There is code here to handle 'CONF.network_manager' being set to
'NeutronManager'. This is a long-dead pattern, having been replaced by
the 'use_neutron' option, and would never be called given that the
'validate_network_plugin' decorator ensures the functions can only be
used for nova-network. Just remove it.

Change-Id: I1dbfe3a0d9ef15176b17c7a959c68c6b8be879fa
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>