Backport tox.ini to switch to stestr
The pike branch was still using ostestr (instead of stestr) which makes
running tests significantly different from queens or master. To make
things behave the same way this commit backports most of the tox.ini
from queens so that pike will behave the same way for running tests.
This does not use the standard backport mechanism because it involves a
lot of different commits over time. It's also not a functional change
for nova itself, so the proper procedure is less important here.

Change-Id: Ie207afaf8defabc1d1eb9332f43a9753a00f784d