Fix server_group_members quota check
For example there are 3 instances in a server group (quota is 5).
When doing multi-creating of 3 more instances in this group
(would have 6 members), current quota checking scheme will fail to
prevent this happening, which is not expected.

This is due to the server_group_members quota check previously
only counting group members that existed as instance records in
cell databases and not accounting for build requests which are
the temporary representation of the instance in the API database
before the instance is scheduled to a cell.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Change-Id: If439f4486b8fe157c436c47aa408608e639a3e15
Closes-Bug: #1780373