Add functional regressions tests for server_group_members OverQuota
Since we started counting quotas in Pike, it is possible to bypass
the server_group_members qouta check if either creating multiple
servers in a single request or creating one server each in multiple
concurrent requests. This is because the server_group_members
count is based on existing server group members in the cell database
and those group members (instances) don't get created in a cell until
we get to conductor and after the scheduler picks a host. In other
words, the server_group_members quota check in the API does not account
for build requests.

Change-Id: Icb268ca2f792bfcefd152ba4c6aa13270d9a7900
Related-Bug: #1780373
(cherry picked from commit f9874e059df50dc81803fcfdfd1045cc09624894)