Fix TypeError in prep_resize allocation cleanup
The CachingScheduler doesn't create allocations in
placement so when prep_resize fails and it tries
to revert allocations, the _revert_allocation() method
returns False and then prep_resize calls
delete_allocation_for_failed_resize() which since change
I7891b98f225f97ad47f189afb9110ef31c810717 has taken a
context argument, but that's not being passed so it
results in a TypeError.

This fixes the TypeError and adds a functional test case
which runs a resize reschedule scenario with the
CachingScheduler.

Change-Id: I0ec0c2d3f7ef8d2274b97a28a175d53921edc9db
Closes-Bug: #1781100