Fix ServerMigrationSampleJsonTestsV2_24 to use its own sample file
The ServerMigrationSampleJsonTestsV2_24 functional tests set up
``self.api.microversion`` to 2.24 in its' setUp(), by doing this
it will not need to add its' own sample files, as there will be
no difference anyway. But this makes Tests that inherent from this
test class impossibe to run as the microversion will always been
overriden.

This patch fix it by change the tests to use self.microversion and
adds its' own sample files.

Part of blueprint abort-live-migration-in-queued-status

Change-Id: I07222efb8d566fa1dd395ef948eae6e6c4b13678