[placement] add error.code on a ConcurrentUpdateDetected
A pending api-sig guideline
I473918ce9c6b49c0b904e93b7140421525f4e7c0 reflects discussion amongst
various parties that adding error codes to structured error responses
should not be considered a breaking API change.

Given that, it is simply "okay" to add them. This change adds
errors.CONCURRENT_UPDATE wherever the ConcurrentUpdateDetected results
in an HTTPConflict (409) response.

In practice this means that if microversion 1.23 or beyond is being
used, there will be a code in any error response, and for some
409 responses that code will be `placement.concurrent_update`.

Note that because we haven't instrumented a way to cause a truly
concurrent update, we've never had real tests for this kind of
thing.

Change-Id: I5a164a0ee298846231113ad7b62b82a3e8f1f409