Add another up-call to the cells v2 caveats list
Due to change I8d426f2635232ffc4b510548a905794ca88d7f99 in Pike,
which ironically was meant to avoid up-calls (I think), it
introduced an up-call during reschedules for server create and
resize to set the instance.availability_zone based on the
alternate host selected during the reschedule.

This adds the up-call to our list of known issues in the cells
v2 docs so we can track the issue and make people aware of it.

Change-Id: Id819f91477613a013b89b1fb0b2def3b0fd4b08c
Related-Bug: #1781286