fix cellv2 delete_host
When trying to delete host that can be found in host_mappings
but not in compute_nodes, current cellv2 delete_host will throw
an exception but does not really handle it.

This patch tries to handle this exception and allow the delete
operation to continue since it shows the host has gone anyway.

Change-Id: I99bd79fb45777edc0e33d846ba478b0a94a1191e
Closes-Bug: #1781391
(cherry picked from commit 0ef4ed96d19a3576d67fdc9c1958610887d57ed4)