Don't overwrite greenthread-local context in host manager
There are a couple of places in host manager where we create a blank
RequestContext for internal database operations without passing
overwrite=False, so they will overwrite the greenthread-local context
that will be used for logging request_id, replacing them with newly
generated request_ids.

This changes the RequestContext creations to get_admin_context calls,
(which uses overwrite=False) to be more explicit about the admin-ness
of the internal database operations we're doing in these methods.

Related-Bug: #1773102

Change-Id: I752eb677d9ccc5ec65147380efe4067456fa312b