Add functional regression test for bug 1778305
This tests the scenario where an operator has upgraded from Ocata =>
Pike but has old, deleted services that share the same hostname with
currently existing, running services. The current behavior is that
ServiceTooOld is raised when we try to generate a UUID for the old,
deleted service while querying for a list of instances. After the bug
is fixed, we should be able to successfully query for a list of
instances.

Related-Bug: #1778305

Change-Id: I1c62a0a5be6c44176b5714af00cc181eb5437585