Add note about reschedules and num_attempts in filter_properties
The "retry" entry in filter_properties is not set if reschedules
are disabled, which happens in these cases:

1. [scheduler]/max_attempts=1
2. The server is forced to a specific host and/or node.

More times than I'd like to admit, I've had to re-learn that
filter_properties['retry']['num_attempts'] will always be >1 in
conductor build_instances during a reschedule because if
reschedules are disabled, the compute service aborts the build
on failure and we don't even get back to conductor.

This change adds a note since it's hard to keep in your head how
the retry logic is all tied together from the API, superconductor,
compute and cell conductor during a reschedule scenario.

Change-Id: I83536b179000f41f9618a4b6f2a16b4440fd61ba
Related-Bug: #1781286