ironic: provide facilities to gracefully navigate versions
For quite some time, the nova virt driver has invoked the ironic
python client library with a static microversion pin which causes
operators to have a failed nova-compute process upon out of order
upgrades, as well as causes ironic's grenade CI jobs to break every
cycle when the pin is changed to a version released in the current
cycle due to established community agreement as well as the grenade
testing framework limitations.

In order to gracefully navigate this, python-ironicclient has
accepted a list of possible versions to negotiate since 2.2.0.

This patch provies a mechanism to validate if we are able to
send a request that was added at a specific version, allowing for
minimal code changes and for logic to only be centered around newer
features being added.

Change-Id: I440689a246538fbc8200687e40480d837b87eb7b
Closes-Bug: #1739440