[placement] Add gabbi coverage for inv of missing rp
There was no gabbi coverage for the case where a GET request was being
made for single class of inventory on a non-existent resource provider.
This adds it.

Note that it would have been possible to add an error code/comment for
the raised exception, but this was not done in this change because there
is another change in progress which is adding the appropriate code.

Change-Id: I1858f4eb617cc81b41610195bb9a33a9059850f7