Fix wonky reqspec handling in conductor.unshelve_instance
Removes the populate_retry call since we don't reschedule
from failed unshelve calls in the compute. That was added
as a partial fix for bug 1400015 but it was never completed.

The to_legacy/from_primitives stuff in here dropped the is_bfv
setting on the request spec, which means we'd have to
recalculate that every time. Instead, if we're given a valid
RequestSpec, use it, otherwise create a fake one and then we'll
heal the RequestSpec.is_bfv field on that one. Plus all of that
missing request spec compat code should get dropped in Stein
anyway (finally).

Related-Bug: #1469179

Change-Id: I49c4e87d15e6fb0fda1b4efd7252bc5ca2066fb4