Remove stevedore extensions server_create method
All the server_create extensions has been merged in
server main controller.

This commit removes the stevedore extensions detection
method from server controller.

This is completion of part-2 of extension merge work.

Remaining part in this BP is to merge the response extensions
which is last part of this work.

Partially implements: blueprint api-extensions-merge-rocky

Change-Id: I077304f4c2d4e50450eb6c6bf4ec427a407bea20