Report 0 root_gb in resource tracker if instance is bfv.
hypervisor local_gb_used still reports usage even with volume-backed
instances after fix for bug 1469179.

This is because of some code in the resource tracker that is not
boot-from-volume aware when reporting disk usage, it just gets the
root_gb off the flavor:

https://github.com/openstack/nova/blob/eb4f65a7951e921b1cd8d05713e144e72f2f254f/nova/compute/resource_tracker.py#L1455

This patch simply check if the instance is bfv before returning
the result. If it's a volume-backed instance and report 0 for root_gb if
that's the case.

Change-Id: I0839470c4bcfb16590a0d87b306d683b059bf8a9
Closes-bug: #1782393