Add shelve/unshelve wrinkle to volume-backed disk func test
This adds a shelve/unshelve scenario to the functional test
which checks that root_gb from the flavor does not show up
in placement for volume-backed servers. Because the shelve
happens after we've cold migrated the server to a new host,
the fake virt driver's finish_migration() method needed to
be implemented to track the instance on the destination host.

Change-Id: Ica456f2512ebe7814c5d20f205ba89b49c42050a
Related-Bug: #1469179