Consider network NUMA affinity for move operations
There are two things required here. Firstly, we need to start
(consistently) storing the physnet and tunneled status of various
networks in 'Instance.info_cache.network_info'. Once we have this
information, we can use it to populate the
'RequestSpec.network_metadata', which can be consumed by later changes
in the series.

Note that live migrations and evacuations with a forced destination
host bypass the scheduler which also means there will be no NUMA
vswitch affinity "claims" on the destination host for those force
move operations. This is not a regression since (1) live migration
is not NUMA affinity aware nor does claims anyway (see blueprint
numa-aware-live-migration) and (2) forced host evacuate already does
not perform claims on the compute since the scheduler is bypassed
so the limits passed to compute, used for the claim, are empty.

Part of blueprint numa-aware-vswitches

Change-Id: I393bd58b8fede38af98ded0c7be099ef22b6f75b